refactor: 拡張機能のディレクトリ構成を整理
- 機能別にディレクトリを整理し、関連ファイルをグループ化
- src/features/ 配下に機能別サブディレクトリを作成:
  - templates/: テンプレート管理機能
  - dependency-graph/: 依存グラフ分析機能
  - api-analysis/: API分析機能
  - graphql/: GraphQL関連機能
  - protobuf/: Protobuf関連機能
- src/utils/ 配下に共通ユーティリティを配置
- 各機能ディレクトリにindex.tsを追加し、公開APIを明確化
- テストファイルもソース構造に合わせて整理
- インポートパスを新しい構造に合わせて更新
- ARCHITECTURE.mdを追加し、新しい構造を文書化

この変更により、コードの保守性と可読性が向上し、
問題発生時の原因特定が容易になります。

Closes #25
