+++ src/features/api-analysis/index.ts
@@ -1,0 +1,3 @@
+export { ApiTestSkeletonGenerator } from './apiTestSkeletonGenerator';
+export { RestApiAnalyzer } from './restApiAnalyzer';
+export type { RestEndpoint } from './restApiAnalyzer';
