+++ src/features/dependency-graph/index.ts
@@ -1,0 +1,4 @@
+export { DependencyGraphAnalyzer } from './dependencyGraphAnalyzer';
+export { DependencyGraphWebview } from './dependencyGraphWebview';
+export { DependencyTreeProvider } from './dependencyTreeProvider';
+export type { DependencyNode, DependencyGraph } from './dependencyGraphAnalyzer';
