+++ src/features/graphql/index.ts
@@ -1,0 +1,3 @@
+export { GraphQLAnalyzer } from './graphqlAnalyzer';
+export { GraphQLDocsGenerator } from './graphqlDocsGenerator';
+export type { GraphQLOperation } from './graphqlAnalyzer';
