+++ src/features/protobuf/index.ts
@@ -1,0 +1,1 @@
+export { ProtobufFieldNumberer } from './protobufFieldNumberer';
