+++ src/features/templates/index.ts
@@ -1,0 +1,2 @@
+export { TemplateManager } from './templateManager';
+export { TemplateEditorProvider } from './templateEditorProvider';
