+++ src/features/templates/templateEditorProvider.ts
