+++ src/features/templates/templateManager.ts
