+++ src/utils/ignorePatternUtils.ts
